test(accounts): run composition lifecycle on desktop
